requirements.txt
httpx>=0.27.0
apscheduler>=3.10.4
python-dotenv>=1.0.0
pytz>=2024.1
structlog>=24.1.0
pydantic>=2.6.0
fastapi>=0.115.0
uvicorn>=0.30.0
anthropic>=0.40.0
